Refactor and optimize
