+++ src/MarkdownEditor/Label.tsx
@@ -2,7 +2,7 @@ import React, {ForwardedRef, forwardRef, useContext} from 'react'
 import {SxProp} from '../sx'
 import InputLabel from '../_InputLabel'
 import {MarkdownEditorSlot} from './MarkdownEditor'
-import {MarkdownEditorContext} from './MarkdownEditorContext'
+import {MarkdownEditorContext} from './_MarkdownEditorContext'
 
 type LabelProps = SxProp & {
   visuallyHidden?: boolean
